+++ basicArithmetic.h
@@ -1,10 +1,11 @@
-class basicArithmetic {
+class basicArithmetic
+{
 private:
   /* data */
 
 public:
-  int sum(int a, int b);
-  int divide(int a, int b);
-  int multiply(int a, int b);
-  int subtract(int a, int b);
+  float sum(float a, float b);
+  float divide(float a, float b);
+  float multiply(float a, float b);
+  float subtract(float a, float b);
 };
